python/packages/awsutils/src/test/tests/requirements.txt
#
####### requirements.txt #######
#
###### Requirements without Version Specifiers ######
html-testRunner
coverage
pyyaml
moto
#
###### Requirements with Version Specifiers ######
#   See https://www.python.org/dev/peps/pep-0440/#version-specifiers
